dbx-listing/pyproject.toml
[project]
name = "dbx-listing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dacite>=1.9.2",
    "databricks-sdk>=0.76.0",
    "jsondiff>=2.2.1",
    "requests>=2.32.5",
]

[project.scripts]
dbx-listing = "dbx_listing.__main__:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
